Patch for issue #12 - (tested in immediate window)

+++ src/uComponents.DataTypes/ImagePoint/ImagePointDataEditor.cs
@@ -269,7 +269,7 @@ namespace uComponents.DataTypes.ImagePoint
                 catch
                 {
                     // node, media or member with specified property couldn't be found
-                    // TODO: if debug mode on, then thow exception, else be silent
+                    // TODO: if debug mode on, then throw exception, else be silent
                 }
 
                 if (!string.IsNullOrWhiteSpace(imageUrl))
@@ -296,12 +296,12 @@ namespace uComponents.DataTypes.ImagePoint
             }            
             else if (this.options.Width > 0 && this.options.Height == 0)
             {
-                // width set, so calulate height
+                // width set, so calculate height
                 width = this.options.Width;
 
                 if (imageSize.Height > 0)
                 {
-                    height = (int)(imageSize.Height / ((decimal)(imageSize.Width / this.options.Width)));
+                    height = (int)(imageSize.Height / ((decimal)imageSize.Width / (decimal)this.options.Width));
                 }
             }            
             else if (this.options.Width == 0 && this.options.Height > 0)
@@ -309,7 +309,7 @@ namespace uComponents.DataTypes.ImagePoint
                 // height set, so calculate width
                 if (imageSize.Width > 0)
                 {
-                    width = (int)(imageSize.Width / ((decimal)(imageSize.Height / this.options.Height)));
+                    width = (int)(imageSize.Width / ((decimal)imageSize.Height / (decimal)this.options.Height));
                 }
 
                 height = this.options.Height;
